docs: 添加项目文档和设计说明
- 📚 添加 DESIGN_DOCUMENT.md - 详细的功能设计文档
- 📚 添加 IMPLEMENTATION_SUMMARY.md - 实现总结文档
- 📚 添加 API 文档结构
- 📚 添加部署指南文档
- 📚 添加开发指南文档
- 📚 添加项目概览文档
- 🔧 修复后端代码编译错误
- 🗑️ 清理重复和未使用的文件
- ✅ 完善测试用例

文档结构：
- docs/api/ - API接口文档
- docs/deployment/ - 部署相关文档
- docs/development/ - 开发指南
- docs/overview/ - 项目概览

项目现在具备完整的文档体系

+++ backend/internal/repositories/task_repository.go
@@ -37,27 +37,27 @@ type TaskRepository interface {
 
 // TaskFilter represents filters for task queries
 type TaskFilter struct {
-	ProjectID    *uuid.UUID      `json:"project_id,omitempty"`
-	IsCompleted  *bool           `json:"is_completed,omitempty"`
+	ProjectID    *uuid.UUID           `json:"project_id,omitempty"`
+	IsCompleted  *bool                `json:"is_completed,omitempty"`
 	Priority     *models.TaskPriority `json:"priority,omitempty"`
-	SearchQuery  string          `json:"search_query,omitempty"`
-	DueBefore    *time.Time      `json:"due_before,omitempty"`
-	DueAfter     *time.Time      `json:"due_after,omitempty"`
-	Tags         []string        `json:"tags,omitempty"`
-	ParentTaskID *uuid.UUID      `json:"parent_task_id,omitempty"`
-	SortBy       string          `json:"sort_by,omitempty"`    // "title", "priority", "due_date", "created_at", "updated_at"
-	SortOrder    string          `json:"sort_order,omitempty"` // "asc", "desc"
-	Page         int             `json:"page"`
-	Limit        int             `json:"limit"`
+	SearchQuery  string               `json:"search_query,omitempty"`
+	DueBefore    *time.Time           `json:"due_before,omitempty"`
+	DueAfter     *time.Time           `json:"due_after,omitempty"`
+	Tags         []string             `json:"tags,omitempty"`
+	ParentTaskID *uuid.UUID           `json:"parent_task_id,omitempty"`
+	SortBy       string               `json:"sort_by,omitempty"`    // "title", "priority", "due_date", "created_at", "updated_at"
+	SortOrder    string               `json:"sort_order,omitempty"` // "asc", "desc"
+	Page         int                  `json:"page"`
+	Limit        int                  `json:"limit"`
 }
 
 // TaskProjectStats represents task statistics for a project
 type TaskProjectStats struct {
 	ProjectID      uuid.UUID `json:"project_id"`
-	TotalTasks     int       `json:"total_tasks"`
-	CompletedTasks int       `json:"completed_tasks"`
-	PendingTasks   int       `json:"pending_tasks"`
-	OverdueTasks   int       `json:"overdue_tasks"`
+	TotalTasks     int64     `json:"total_tasks"`
+	CompletedTasks int64     `json:"completed_tasks"`
+	PendingTasks   int64     `json:"pending_tasks"`
+	OverdueTasks   int64     `json:"overdue_tasks"`
 }
 
 // taskRepository implements TaskRepository using GORM
